Use single quotes for building the go test errorformat
This way the quotes don't have to be doubled up ("\\t" -> '\t'). Also
use explicit scoping (format -> l:format) and strict comparison (== ->
is) as a matter of good practices, while I'm here.

The true reason for this patch is selfish: so I can more easily copy it
to a "gotest.vim" compiler.